Bugfix for compilation with x86_64-pc-windows-gnu
This commit fixes the existing issues preventing compilation of zmq-sys
when compiling for x86_64-pc-windows-gnu. This fixes the flag issues
referenced by [this
issue](https://github.com/jean-airoldie/zeromq-src-rs/issues/49) (even
though the original bug reporter stopped responding to your replies).

In addition to the flag changes, in order for zmq-sys to compile for
x86_64-pc-windows-gnu the windows networking file `wepoll.c` has to be
compiled with gcc instead of g++. It looks like an attempt was made for
this when setting `wepoll.c` to be compiled as a c source, but without
`build.cpp(false)` the c file still gets compiled with g++, so I added
to both the `add_cpp_sources()` and `add_c_sources()` fn instructions to
set the compiler instructions for each file. This expresses clearer
intent and because instruction overhead isn't really a huge concern
here.

This was tested by compiling on both linux (6.13.7-arch1-1) and
windows-x86_64-pc-windows-gnu, and succeeding in compiling on both. I
don't have a lot of experience doing this, so feedback is welcome.